refactor: coloque o GerarBannerCliente na pasta util da feature
