+++ frontend/src/features/clients/utils/getGerarBannerCliente.ts
